refactor(parse): move the smcat parser into its own folder
