+++ .prettierignore
@@ -1,5 +1,5 @@
 src/cli/attributes-parser.js
-src/parse/smcat-parser.js
+src/parse/smcat/smcat-parser.js
 src/render/dot/dot.states.template.js
 src/render/dot/dot.template.js
 src/render/html/html.template.js
