+++ test/render/smcat.spec.js
@@ -1,6 +1,6 @@
 const expect = require("chai").expect;
 const convert = require("../../src/render/smcat");
-const parser = require("../../src/parse/smcat-parser");
+const parser = require("../../src/parse/smcat/smcat-parser");
 
 const programASTPairs = require("../parse/00-no-transitions.json")
   .concat(require("../parse/01-transitions-only.json"))
